Authorize companies

+++ backend/src/main/java/com/nbu/logisticcompany/services/CompanyServiceImpl.java
@@ -1,10 +1,12 @@
 package com.nbu.logisticcompany.services;
 
 import com.nbu.logisticcompany.entities.Company;
+import com.nbu.logisticcompany.entities.Role;
 import com.nbu.logisticcompany.entities.User;
 import com.nbu.logisticcompany.entities.dtos.company.CompanyOutDto;
 import com.nbu.logisticcompany.exceptions.DuplicateEntityException;
 import com.nbu.logisticcompany.exceptions.EntityNotFoundException;
+import com.nbu.logisticcompany.exceptions.UnauthorizedOperationException;
 import com.nbu.logisticcompany.repositories.interfaces.CompanyRepository;
 import com.nbu.logisticcompany.services.interfaces.CompanyService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,10 @@ import java.util.Optional;
 
 @Service
 public class CompanyServiceImpl implements CompanyService {
+
+    private static final String UNAUTHORIZED_CREATE = "Only Admins can create companies";
+    private static final String UNAUTHORIZED_UPDATE = "Only Admins can update companies";
+    private static final String UNAUTHORIZED_DELETE = "Only Admins can delete companies";
 
     private final CompanyRepository companyRepository;
 
@@ -31,7 +37,7 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Company getByName(String name) {
-        return companyRepository.getByField("name",name);
+        return companyRepository.getByField("name", name);
     }
 
     @Override
@@ -40,12 +46,15 @@ public class CompanyServiceImpl implements CompanyService {
     }
 
     @Override
-    public List<CompanyOutDto> getCompanyIncome(int companyId, LocalDateTime periodStart, LocalDateTime periodEnd){
+    public List<CompanyOutDto> getCompanyIncome(int companyId, LocalDateTime periodStart, LocalDateTime periodEnd) {
         return companyRepository.getCompanyIncome(companyId, periodStart, periodEnd);
     }
 
     @Override
-    public void create(Company company) {
+    public void create(Company company, User creator) {
+        if (!creator.getRoles().contains(Role.ADMIN)) {
+            throw new UnauthorizedOperationException(UNAUTHORIZED_CREATE);
+        }
         boolean duplicateCompany = true;
         try {
             companyRepository.getByField("name", company.getName());
@@ -60,17 +69,17 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public void update(Company companyToUpdate, User user) {
-       /* if (companyToUpdate.getId() != user.getId()) {
-            //throw new UnauthorizedOperationException(UNAUTHORIZED_UPDATE);
-        }*/
+        if (!user.getRoles().contains(Role.ADMIN)) {
+            throw new UnauthorizedOperationException(UNAUTHORIZED_UPDATE);
+        }
         companyRepository.update(companyToUpdate);
     }
 
     @Override
     public void delete(int companyId, User user) {
-       /* if (user.getId() != companyId) {
-            //throw new UnauthorizedOperationException(UNAUTHORIZED_DELETE);
-        }*/
+        if (!user.getRoles().contains(Role.ADMIN)) {
+            throw new UnauthorizedOperationException(UNAUTHORIZED_DELETE);
+        }
         companyRepository.delete(companyId);
     }
